refactor: Core modules organized in folders.
