+++ src/core/graphics/CRTEffect.ts
