+++ src/core/graphics/GraphicsDetailMode.ts
