+++ src/core/graphics/gradient.ts
