+++ src/core/graphics/grayscale.ts
