+++ src/core/math/Area.ts
