+++ src/core/math/easings.ts
