+++ src/core/platform/deviceDetection.ts
