+++ src/window.ts
@@ -23,7 +23,7 @@
  */
 
 import { canvas } from "./graphics";
-import { VectorMutable } from "./core/Vector";
+import { VectorMutable } from "./core/math/Vector";
 
 let canvasScale = 1;
